api/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Database
aiosqlite==0.19.0

# AI integration (for image ingestion and analysis)
anthropic>=0.40.0

# Image processing
Pillow==10.1.0
python-magic==0.4.27

# EXIF extraction
piexif==1.1.3

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP client (for external APIs)
httpx==0.25.2

# Utilities
python-slugify==8.0.1
